coffee_back_end/requirements.txt
django>=5.2.0,<6.0.0
djangorestframework>=3.16.0,<4.0.0
django-cors-headers>=4.3.0,<5.0.0
Pillow>=10.0.0,<11.0.0  # 用于图像处理
django-filter>=23.0,<24.0  # 用于过滤
django-extensions>=3.2.0,<4.0.0  # 提供额外的管理命令
djangorestframework-simplejwt>=5.2.0,<6.0.0  # JWT认证
mysqlclient>=2.1.0  # MySQL连接库 